requirements.txt
alembic==1.7.7
attrs==21.4.0
black==22.1.0
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
ConfigArgParse==1.5.3
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-mongoengine==1.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gevent==21.12.0
geventhttpclient==1.5.3
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
locust==2.8.3
Mako==1.2.0
MarkupSafe==2.1.0
mock==4.0.3
mongoengine==0.24.0
mongomock==4.0.0
msgpack==1.0.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
psutil==5.9.0
py==1.11.0
pymongo==4.0
pyparsing==3.0.7
pytest==7.1.1
python-dotenv==0.19.2
pyzmq==22.3.0
requests==2.27.1
roundrobin==0.0.2
sentinels==1.0.0
six==1.16.0
slack-logger==0.3.1
SQLAlchemy==1.4.33
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
WTForms==3.0.1
zope.event==4.5.0
zope.interface==5.4.0
